[장준영] 리뷰작성 수정중...

+++ booking_c/src/main/java/goodee/gdj58/booking_c/vo/ReviewImg.java
@@ -8,5 +8,5 @@ public class ReviewImg {
 	private String reviewImgSaveName;
 	private String reviewImgOriginName;
 	private String reviewImgKind;
-	private int reviewImgSize;
+	private long reviewImgSize;
 }
